requirements.txt
Flask>=2.3.0
pandas>=1.5.0
scikit-learn>=1.2.0
PySastrawi>=1.2.0
rank-bm25>=0.2.2
requests>=2.28.0
feedparser>=6.0.0
python-dotenv>=1.0.0
numpy>=1.23.0
joblib>=1.2.0
